Add image quality to result

+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/util/ColorUtil.java
@@ -127,7 +127,7 @@ public final class ColorUtil {
             Timber.e(e);
         }
 
-        return new ColorInfo(commonColor, quality);
+        return new ColorInfo(commonColor, (int) quality);
     }
 
     /**
